refactor: entities model/types 분리
